render remove button if movie already exist inside user's movie stack
